+++ src/config/values/mod.rs
@@ -359,7 +359,7 @@ macro_rules! numbers {
     };
 }
 
-numbers!(u8, u16, u32, i8, i16, i32);
+numbers!(u8, u16, u32, i8, i16, i32, i64);
 
 impl From<usize> for Value {
     fn from(n: usize) -> Self {